+++ cycles-admin-service/cycles-admin-service-data/src/test/java/io/runcycles/admin/data/integration/RedisIntegrationTest.java
@@ -99,7 +99,7 @@ class RedisIntegrationTest {
         webhookRepository = new WebhookRepository();
         injectField(webhookRepository, "jedisPool", jedisPool);
         injectField(webhookRepository, "objectMapper", objectMapper);
-        injectField(webhookRepository, "cryptoService", new CryptoService(""));
+        injectField(webhookRepository, "cryptoService", new CryptoService("", true));
 
         tenantCloseWorkRepository = new TenantCloseWorkRepository(jedisPool, objectMapper);
         idempotencyStore = new IdempotencyStore(jedisPool, objectMapper);